backend/requirements.txt
# File: requirements.txt

# Framework utama untuk API
fastapi==0.110.3

# Server untuk menjalankan FastAPI
uvicorn[standard]==0.29.0

# Interaksi dengan database
sqlalchemy==2.0.41

# Hashing password
passlib[bcrypt]==1.7.4

# JSON Web Tokens (JWT) untuk autentikasi
python-jose[cryptography]==3.3.0

# Dibutuhkan oleh FastAPI untuk menangani form login
python-multipart==0.0.9

# Untuk membuat request ke API lain (OpenRouter.ai)
httpx==0.27.0

# Manajemen konfigurasi dari environment variables (BARU)
pydantic-settings==2.2.1

# Driver untuk database PostgreSQL (untuk persiapan produksi)
psycopg2-binary==2.9.9